Improved code structure
